+++ common/src/main/java/com/jaquadro/minecraft/storagedrawers/client/model/DecoratedDrawerModel.java
@@ -93,6 +93,16 @@ public abstract class DecoratedDrawerModel implements BakedModel
         else if (isClaimed)
             emitModel.accept(DrawerModelStore.getModel(DrawerModelStore.DynamicPart.CLAIM, dir, half));
 
+        DrawerModelStore.DynamicPart priorityPart = switch (attr.getPriority()) {
+            case 1 -> DrawerModelStore.DynamicPart.PRIORITY_P1;
+            case 2 -> DrawerModelStore.DynamicPart.PRIORITY_P2;
+            case -1 -> DrawerModelStore.DynamicPart.PRIORITY_N1;
+            case -2 -> DrawerModelStore.DynamicPart.PRIORITY_N2;
+            default -> null;
+        };
+        if (priorityPart != null)
+            emitModel.accept(DrawerModelStore.getModel(priorityPart, dir, half));
+
         if (attr.isVoid())
             emitModel.accept(DrawerModelStore.getModel(DrawerModelStore.DynamicPart.VOID, dir, half));
         if (attr.isConcealed())